requirements.txt
Django==1.6.11
South==1.0.2
django-moderation==0.3.6
pytz==2015.6
